% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bimaxwellian.R
\docType{data}
\name{bimaxwellian_expr}
\alias{bimaxwellian_expr}
\title{bimaxwellian_expr}
\format{An object of class \code{expression} of length 1.}
\usage{
bimaxwellian_expr
}
\description{
expression for a bimaxwellian momentum distribution in
  cylindrical coordinates
}
\keyword{datasets}
