% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bimaxwellian.R
\name{bimaxwellian_func}
\alias{bimaxwellian_func}
\title{bimaxwellian_func}
\usage{
bimaxwellian_func(p_perp, p_par, n, p_tperp, p_tpar, p_drift)
}
\arguments{
\item{p_perp}{\code{numeric} value of perpendicular momentum component}

\item{p_par}{\code{numeric} value of parallel momentum component}

\item{n}{\code{numeric} particle density.}

\item{p_tperp}{\code{numeric} themal momentum perpendicular to magnetic field}

\item{p_tpar}{\code{numeric} themal momentum parallel to magnetic field}

\item{p_drift}{\code{numeric} drift momentum parallel to magnetic field}
}
\value{
\code{numeric} value of momentum distribution at (p_perp, p_par)
}
\description{
function to evaluate a bimaxwellian momentum distribution
}
