% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bimaxwellian.R
\name{bimaxwellian_grad}
\alias{bimaxwellian_grad}
\title{bimaxwellian_grad}
\usage{
bimaxwellian_grad(p_perp, p_par, n, p_tperp, p_tpar, p_drift)
}
\arguments{
\item{p_perp}{\code{numeric} value of perpendicular momentum component}

\item{p_par}{\code{numeric} value of parallel momentum component}

\item{n}{\code{numeric} particle density.}

\item{p_tperp}{\code{numeric} thermal momentum perpendicular to magnetic field}

\item{p_tpar}{\code{numeric} thermal momentum parallel to magnetic field}

\item{p_drift}{\code{numeric} drift momentum parallel to magnetic field}
}
\value{
\code{list}
}
\description{
function to calculate the gradient of a  bimaxwellian
  momentum distribution with respect to parallel and perpencicular momentum
}
