% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bimaxwellian.R
\name{bimaxwellian_setup}
\alias{bimaxwellian_setup}
\title{bimaxwellian_setup}
\usage{
bimaxwellian_setup(n, T_eV_perp, T_eV_par, v_drift, A)
}
\arguments{
\item{n}{\code{numeric} Particle density.}

\item{A}{\code{numeric} Particle mass number}

\item{T_eV}{\code{numeric} Temperaure in eV.}
}
\value{
\code{list} with momentum distribution setup
}
\description{
Function to return a simple maxwellian momentum distribution
}
