% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilkie_sd.R
\name{birth_velocity}
\alias{birth_velocity}
\title{birth_velocity}
\usage{
birth_velocity(birth_energy, m)
}
\arguments{
\item{birth_energy}{\code{numeric} birth energy in eV.}

\item{m}{\code{numeric} particle mass in kg.}
}
\value{
\code{numeric} birth velocity
}
\description{
Function to calculate the birth velocity of a perticle given its
  mass and birth energy.
}
