% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvtnorm.R
\docType{data}
\name{bvtnorm_expr}
\alias{bvtnorm_expr}
\title{bvtnorm_expr}
\format{
An object of class \code{expression} of length 1.
}
\usage{
bvtnorm_expr
}
\description{
expression for a bivariate normal momentum distribution in
  cylindrical coordinates
}
\keyword{datasets}
