% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvtnorm.R
\name{bvtnorm_func}
\alias{bvtnorm_func}
\title{bvtnorm_func}
\usage{
bvtnorm_func(p_perp, p_par, n, center, covariance, K)
}
\arguments{
\item{p_perp}{\code{numeric} value of perpendicular momentum component}

\item{p_par}{\code{numeric} value of parallel momentum component}

\item{n}{\code{numeric} particle density.}

\item{center}{\code{numeric} center of distribution (2d vector, c(p_perp,
p_par))}

\item{covariance}{\code{numeric} covariance of distribution (2 by 2 matrix)}

\item{K}{\code{numeric} normalizations constant}
}
\value{
\code{numeric} value of momentum distribution at (p_perp, p_par)
}
\description{
function to evaluate a bivariate normal momentum distribution
}
