% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvtnorm.R
\name{bvtnorm_setup}
\alias{bvtnorm_setup}
\title{bvtnorm_setup}
\usage{
bvtnorm_setup(n, center, covariance, A, Z, name)
}
\arguments{
\item{n}{\code{numeric} Particle density.'}

\item{center}{\code{numeric} center of velocity distribution (2d vector)}

\item{covariance}{\code{numeric} covariance of velocity distribution (2 by 2 matrix)}

\item{A}{\code{numeric} Particle mass number}

\item{Z}{\code{numeric} Particle charge number}

\item{name}{\code{character} Name of distribution/particle}
}
\value{
\code{list} with momentum distribution setup
}
\description{
Function to return a bivariate normal momentum distribution
}
\details{
#' @title bvtnorm_grad
#'
#' @description function to calculate the gradient of a a bivariate normal
#'   momentum distribution with respect to parallel and perpencicular momentum
#'
#' @param p_perp \code{numeric} value of perpendicular momentum component
#' @param p_par \code{numeric} value of parallel momentum component
#' @param n \code{numeric} particle density.
#' @param center \code{numeric} center of distribution (2d vector)
#' @param covariance \code{numeric} covariance of distribution (2 by 2 matrix)
#'
#' @return \code{list}
#'
bvtnorm_grad <- deriv(
  expr = bvtnorm_expr,
  namevec = c("p_perp","p_par"),
  function.arg = c("p_perp", "p_par", "n", "center", "covariance")
)
}
