% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_distribution_data_frame.R
\name{calculate_distribution_data_frame}
\alias{calculate_distribution_data_frame}
\title{calculate_distribution_data_frame}
\usage{
calculate_distribution_data_frame(particles, v_par, v_perp)
}
\arguments{
\item{particles}{\code{list} list of particles with distribution setups.}

\item{v_par}{\code{numeric} value of parallel velocity}

\item{v_perp}{\code{numeric} value of perpendicular velocity}

\item{logscale}{\code{logical} if TRUE sets log scale on y-axis}
}
\value{
\code{data.frame} with evaluated distribution in column
  distribution_value
}
\description{
function to evaluate velocity distributions and return in a data
  frame
}
