% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl.R
\name{cl}
\alias{cl}
\title{cl.}
\usage{
cl(l, direction, k_perp, k_par, v_perp, v_par, omega_c)
}
\arguments{
\item{l}{\code{numeric} vector of l-values (Index of Bessel functions).}

\item{direction}{\code{character} spatial direction: x, y or z.}

\item{k_perp}{\code{numeric} length of component of wavevector perpendicular
to the magnetic field.}

\item{k_par}{\code{numeric} length of component of wavevector parallel to the
magnetic field.}

\item{v_perp}{\code{numeric} particle velocity component perpendicular to the
magnetic field.}

\item{v_par}{\code{numeric} particle velocity component  parallel to the
magnetic field.}

\item{omega_c}{\code{numeric} particle angular cyclotron frequency (omega_c =
qB/m).}
}
\value{
\code{complex} value of cl term
}
\description{
Function to calculate cl terms, eq. 26 of Bindslev 1996, JATP
  58, 983.
}
