% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j0j0_element.R
\name{create_integral_settings}
\alias{create_integral_settings}
\title{create_integral_settings}
\usage{
create_integral_settings(
  directions,
  k_perp,
  k_par,
  omega,
  omega_c,
  mass,
  distribution,
  integration_method
)
}
\arguments{
\item{directions}{\code{character} vector with two spatial directions: a
combination of x, y, and z. e.g  xx, xy or yz.}

\item{k_perp}{\code{numeric} length of component of wavevector perpendicular
to the magnetic field.}

\item{k_par}{\code{numeric} length of component of wavevector parallel to the
magnetic field.}

\item{omega}{\code{numeric} angular cyclotron frequency of current
fluctuations.}

\item{omega_c}{\code{numeric} particle angular cyclotron frequency (omega_c =
qB/m).}

\item{mass}{\code{numeric} particle mass in kg.}

\item{distribution}{\code{list} with the velocity distribution and associated
settings.}

\item{integration_method}{\code{character} method to use for integration. One
of "stats", "hcubature", "pcubature", "cuhre", "divonne", "suave", "vegas"}
}
\value{
\code{list} with settinngs for the integral
}
\description{
creates a list of arguments for the integral
}
