% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilkie_sd.R
\name{critical_velocity}
\alias{critical_velocity}
\title{critical_velocity}
\usage{
critical_velocity(n_e, T_e_eV, ions)
}
\arguments{
\item{n_e}{\code{numeric} Electron density.}

\item{T_e_eV}{\code{numeric} Electron temperaure in eV.}

\item{ions}{\code{data frame} with information on all ion species,
containing columns "n" (ion density), "A" (ion mass number), and "Z"
(ion charge number)}
}
\value{
\code{numeric} critical velocity
}
\description{
Function to calculate the criticl velocity, eq. 2.1 of Wilkie
  2018, https://arxiv.org/abs/1808.01934v2.
}
