% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilkie_sd.R
\name{debye_length}
\alias{debye_length}
\title{debye_length}
\usage{
debye_length(n_e, T_i_eV, T_e_eV)
}
\arguments{
\item{n_e}{\code{numeric} electron density.}

\item{T_i_eV}{\code{numeric} Ion temperaure in eV.}

\item{T_e_eV}{\code{numeric} Electron temperaure in eV.}
}
\value{
\code{numeric} Debye length
}
\description{
Function to calculate the debye length (eq. 1.5 of Swanson 2008)
}
