% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dist_2D_examples}
\alias{dist_2D_examples}
\title{Examples of 2D velocity distributions with variation in both v_par and
 v_perp for a number of velocity distributions:}
\format{
A data frame with 630000 rows and 8 variables:
\describe{
  \item{name}{Distribution name}
  \item{v_par}{Parallel velocity}
  \item{v_perp}{Perpendiculer velocity}
  \item{p_scale}{Momentum scale - used to scale momentum in integrals, so it is near 1}
  \item{A}{mass number}
  \item{p_par}{parallel momentum}
  \item{p_perp}{perpendicular momentum}
  \item{value}{momentum distribution value (i.e. density)}
  }
}
\source{
File /data/raw/j0j0_examples
}
\usage{
dist_2D_examples
}
\description{
* maxwellian
 * bimaxwellian
 * lorentzian
 * ring
 * bvtnorm
 * slowdown_b5
 * slowdown_b0
}
\keyword{datasets}
