% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{distribution_examples}
\alias{distribution_examples}
\title{Examples 1D velocity distributions at v_perp = 0}
\format{
A data frame with 7000 rows and 8 variables:
\describe{
  \item{name}{Distribution name}
  \item{v_par}{Parallel velocity}
  \item{v_perp}{Perpendiculer velocity}
  \item{p_scale}{Momentum scale - used to scale momentum in integrals, so it is near 1}
  \item{A}{mass number}
  \item{p_par}{parallel momentum}
  \item{p_perp}{perpendicular momentum}
  \item{value}{momentum distribution value (i.e. density)}
  }
}
\source{
File /data/raw/j0j0_examples
}
\usage{
distribution_examples
}
\description{
* maxwellian
 * ring
 * slowdown_b5
}
\keyword{datasets}
