% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_tools.R
\name{eval_distribution}
\alias{eval_distribution}
\title{eval_distribution}
\usage{
eval_distribution(distribution, p_perp, p_par)
}
\arguments{
\item{distribution}{\code{list} with the momentum distribution.}

\item{p_perp}{\code{numeric} value of perpendicular momentum component.}

\item{p_par}{\code{numeric} value of parallel momentum component.}
}
\value{
\code{numeric} value of distribution at (p_perp, p_par).
}
\description{
evaluate a distribution at specific values of p_perp and p_par
}
