% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_tools.R
\name{eval_distribution_list}
\alias{eval_distribution_list}
\title{eval_distribution_list}
\usage{
eval_distribution_list(distribution_list, p_perp, p_par)
}
\arguments{
\item{distribution_list}{\code{list} with a list of momentum distributions.}

\item{p_perp}{\code{numeric} value of perpendicular momentum component.}

\item{p_par}{\code{numeric} value of parallel momentum component.}
}
\value{
\code{numeric} sum of values of distributions at (p_perp, p_par).
}
\description{
evaluate a list of distributions at specific values of p_perp
  and p_par
}
