% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_tools.R
\name{eval_homogeneous_distribution}
\alias{eval_homogeneous_distribution}
\title{eval_homogeneous_distribution}
\usage{
eval_homogeneous_distribution(distribution, p)
}
\arguments{
\item{distribution}{\code{list} with the momentum distribution.}

\item{p}{\code{numeric} length of momentum vector.}
}
\value{
\code{numeric} value of distribution at p.
}
\description{
evaluate a homogeneous distribution at a specific values of p
}
