% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j0j0_element.R
\name{evaluate_integral}
\alias{evaluate_integral}
\title{create_integral_settings}
\usage{
evaluate_integral(integral_settings, integration_method)
}
\arguments{
\item{integral_settings}{list of settings for the integral}

\item{integration_method}{name of integration method}
}
\value{
\code{numeric} value of integral
}
\description{
creates a list of arguments for the integral
}
