% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j0j0_integrand.R
\name{expand_l_values}
\alias{expand_l_values}
\title{expand_l_values}
\usage{
expand_l_values(l_values, l_values_done, updown, nl)
}
\arguments{
\item{l_values}{\code{numeric} vector of l-values to expand.}

\item{l_values_done}{\code{numeric} vector of l-values already done.}

\item{updown}{\code{character} expand "up" or "down".}

\item{nl}{\code{numeric} number of l-values to expand with}
}
\value{
\code{numeric}
}
\description{
expand a vector of l-values up or down by nl
}
