% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilkie_sd.R
\name{fast_ion_slowdown_time}
\alias{fast_ion_slowdown_time}
\title{fast_ion_slowdown_time}
\usage{
fast_ion_slowdown_time(n_e, T_i_eV, T_e_eV, A, Z)
}
\arguments{
\item{n_e}{\code{numeric} electron density.}

\item{T_i_eV}{\code{numeric} Ion temperaure in eV.}

\item{T_e_eV}{\code{numeric} Electron temperaure in eV.}

\item{A}{\code{numeric} Ion mass number}

\item{Z}{\code{numeric} Ion charge number}
}
\value{
\code{numeric} plasma parameter
}
\description{
Function to calculate fast ion slowdown time, eq. 2.2 of Wilkie
  2018, https://arxiv.org/abs/1808.01934v2
}
