% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalized_lorentzian.R
\name{find_glz_theta}
\alias{find_glz_theta}
\title{find_glz_theta}
\usage{
find_glz_theta(T_eV, kp, A)
}
\arguments{
\item{T_eV}{\code{numeric} Temperature in eV.}

\item{kp}{\code{numeric} kappa parameter (spectral index).}

\item{A}{\code{numeric} Particle mass number}
}
\value{
\code{numeric} theta parameter
}
\description{
Function to calculate the theta parameter in a generalized Lorentzian
}
