% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_maxwellian.R
\name{find_p_term}
\alias{find_p_term}
\title{find_p_term}
\usage{
find_p_term(T_eV, A)
}
\arguments{
\item{T_eV}{\code{numeric} Temperaure in eV.}

\item{A}{\code{numeric} Particle mass number}
}
\value{
\code{numeric} thermal momentum
}
\description{
Function to calculate thermal momentum for a simple maxwellian
}
