% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalized_lorentzian.R
\docType{data}
\name{generalized_lorentzian_expr}
\alias{generalized_lorentzian_expr}
\title{generalized_lorentzian_expr}
\format{
An object of class \code{expression} of length 1.
}
\usage{
generalized_lorentzian_expr
}
\description{
expression for a generalized Lorentzian (kappa) distribution,
  see
  https://www.spenvis.oma.be/help/background/distributions/distributions.html.
}
\keyword{datasets}
