% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalized_lorentzian.R
\name{generalized_lorentzian_func}
\alias{generalized_lorentzian_func}
\title{generalized_lorentzian_func}
\usage{
generalized_lorentzian_func(p_perp, p_par, n, theta, kp)
}
\arguments{
\item{p_perp}{\code{numeric} value of perpendicular momentum component}

\item{p_par}{\code{numeric} value of parallel momentum component}

\item{n}{\code{numeric} particle density.}

\item{theta}{\code{numeric} Particle themal momentum parameter}

\item{kp}{\code{numeric} kappa parameter (spectral index). (kappa is a
function in base R so here we use kp for kappa)}
}
\value{
\code{numeric} value of momentum distribution at (p_perp, p_par)
}
\description{
function to evaluate a generalized Lorentzian distribution
}
