% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalized_lorentzian.R
\name{generalized_lorentzian_grad}
\alias{generalized_lorentzian_grad}
\title{generalized_lorentzian_grad}
\usage{
generalized_lorentzian_grad(p_perp, p_par, n, theta, kp)
}
\arguments{
\item{p_perp}{\code{numeric} value of perpendicular momentum component}

\item{p_par}{\code{numeric} value of parallel momentum component}

\item{n}{\code{numeric} particle density.}

\item{theta}{\code{numeric} Particle themal momentum parameter}

\item{kp}{\code{numeric} kappa parameter (spectral index).}
}
\value{
\code{list}
}
\description{
function to calculate the gradient of a generalized Lorentzian
  distribution momentum distribution with respect to parallel and
  perpendicular momentum
}
