% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalized_lorentzian.R
\name{generalized_lorentzian_setup}
\alias{generalized_lorentzian_setup}
\title{generalized_lorentzian_setup}
\usage{
generalized_lorentzian_setup(n, T_eV, kp, A)
}
\arguments{
\item{n}{\code{numeric} Particle density.}

\item{T_eV}{\code{numeric} Temperaure in eV.}

\item{kp}{\code{numeric} kappa parameter (spectral index).}

\item{A}{\code{numeric} Particle mass number}
}
\value{
\code{list} with momentum distribution setup
}
\description{
Function to return a simple maxwellian momentum distribution
}
