% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j0j0.R
\name{initiate_spec}
\alias{initiate_spec}
\title{initiate_spec}
\usage{
initiate_spec(
  k,
  phi,
  frequencies,
  directions,
  B,
  particles,
  integration_method
)
}
\arguments{
\item{k}{\code{numeric} length of the fluctuation wavevector.}

\item{phi}{\code{numeric} angle (in degrees) between the magnetic field and
the fluctuation wavevector}

\item{frequencies}{\code{numeric} fluctuation frequencies in Hz.}

\item{directions}{\code{character} one or more spatial directions ("x", "y"
or "z").}

\item{B}{\code{numeric} strength of magnetic field in Tesla.}

\item{particles}{\code{list} with mass, charge, and momentum distribution.}

\item{integration_method}{\code{character} method to use for integration. One
of "stats", "hcubature", "pcubature", "cuhre", "divonne", "suave", "vegas"}
}
\value{
\code{data.frame}
}
\description{
creates a \code{data.frame}, spec, with all combinations of the
  values in the inpurt parameters
}
