% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_tools.R
\name{integrand_distribution}
\alias{integrand_distribution}
\title{integrand_distribution}
\usage{
integrand_distribution(p, distribution)
}
\arguments{
\item{p}{\code{numeric} momentum vector: p = c(p_perp, p_par). Of order 1.}

\item{distribution}{\code{list} with the momentum distribution.}
}
\value{
\code{numeric} value of integrand at p = c(p_perp, p_par)
}
\description{
function to calculate the integrand in an integral of a
  distribution over all momentum space. The integral is carried out in
  cylindrical coordinates, so the volume element in the integral is p_perp *
  dp_perp * dp_par. Therefore the distribution is here multiplied by p_perp.
  Also note the input p must be of order 1 and is scaled in the function by
  distribution[["p_scale"]].
}
