% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_tools.R
\name{integrand_homogeneous_distribution}
\alias{integrand_homogeneous_distribution}
\title{integrand_homogeneous_distribution}
\usage{
integrand_homogeneous_distribution(p, distribution)
}
\arguments{
\item{p}{\code{numeric} length ofmomentum vector: p = norm(c(p_perp, p_par)).
Of order 1.}

\item{distribution}{\code{list} with the momentum distribution.}
}
\value{
\code{numeric} value of integrand at p
}
\description{
function to calculate the integrand in an integral of a
  distribution over all momentum space. The integral is carried out in
  spherical coordinates. Therefore the distribution is here multiplied by
  p^2. Also note the input p must be of order 1 and is scaled in the function
  by distribution[["p_scale"]].
}
