% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_tools.R
\name{integrate_distribution}
\alias{integrate_distribution}
\title{integrate_distribution}
\usage{
integrate_distribution(distribution)
}
\arguments{
\item{distribution}{\code{list} with the momentum distribution.}
}
\value{
\code{numeric} value of integral
}
\description{
function to calculate integral of a
  distribution over all momentum space. The integral is carried out in
  cylindrical coordinates.
}
