% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_tools.R
\name{integrate_distribution_list}
\alias{integrate_distribution_list}
\title{integrate_distribution_list}
\usage{
integrate_distribution_list(distribution_list)
}
\arguments{
\item{distribution_list}{\code{list} wit a list of momentum distributions.}
}
\value{
\code{numeric}
}
\description{
function to calculate integral of a summed list of
  distributions over all momentum space. The integral is carried out in
  cylindrical coordinates.
}
