% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_tools.R
\name{integrate_homogeneous_distribution}
\alias{integrate_homogeneous_distribution}
\title{integrate_homogeneous_distribution}
\usage{
integrate_homogeneous_distribution(distribution, limits = c(0, Inf))
}
\arguments{
\item{distribution}{\code{list} with the momentum distribution.}
}
\value{
\code{numeric} value of integral
}
\description{
function to calculate integral of a homogeneous
  distribution over all momentum space. The integral is carried out in
  spherical coordinates.
}
