% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j0j0.R
\name{j0j0}
\alias{j0j0}
\title{j0j0}
\usage{
j0j0(k, phi, frequencies, directions = c("x", "y", "z"), B, particles,
  integration_method)
}
\arguments{
\item{k}{\code{numeric} length of the fluctuation wavevector.}

\item{phi}{\code{numeric} angle (in degrees) between the magnetic field and
the fluctuation wavevector}

\item{frequencies}{\code{numeric} fluctuation frequencies in Hz.}

\item{directions}{\code{character} one or more spatial directions ("x", "y"
or "z").}

\item{B}{\code{numeric} strength of magnetic field in Tesla.}

\item{particles}{\code{list} with mass, charge, and momentum distribution.}

\item{integration_method}{\code{character} method to use for integration. One
of "stats", "hcubature", "pcubature", "cuhre", "divonne", "suave", "vegas"}
}
\value{
\code{data.frame}
}
\description{
Wrapper function for j0j0_element to calculate a spectrum of
  j0j0 values for multiple frequencies and directions. Parallelized with
  future/furrr.
}
