% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j0j0_element.R
\name{j0j0_element}
\alias{j0j0_element}
\title{j0j0_element}
\usage{
j0j0_element(directions, k, phi, frequency, B, A, Z, distribution)
}
\arguments{
\item{directions}{\code{character} vector with two spatial directions: a
combination of x, y, and z. e.g  xx, xy or yz.}

\item{k}{\code{numeric} length of the fluctuation wavevector.}

\item{phi}{\code{numeric} angle (in degrees) between the magnetic field and
the fluctuation wavevector}

\item{frequency}{\code{numeric} fluctuation frequency in Hz.}

\item{B}{\code{numeric} strength of magnetic field in Tesla.}

\item{A}{\code{numeric} particle mass number.}

\item{Z}{\code{numeric} particle charge number.}

\item{distribution}{\code{list} with the velocity distribution and associated
settings.}
}
\value{
\code{complex}
}
\description{
Function to calculate terms in the  unscreened current
  fluctuation correlation tensor. I.e th integral in in eq. 25 of Bindslev
  1996, JATP 58, 983.
}
