% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j0j0_spectrum2.R
\name{j0j0_spectrum2}
\alias{j0j0_spectrum2}
\title{j0j0_spectrum2}
\usage{
j0j0_spectrum2(k, phi, frequencies, directions = c("x", "y", "z"), B,
  particles)
}
\arguments{
\item{k}{\code{numeric} length of the fluctuation wavevector.}

\item{phi}{\code{numeric} angle (in degrees) between the magnetic field and
the fluctuation wavevector}

\item{frequencies}{\code{numeric} fluctuation frequencies in Hz.}

\item{directions}{\code{character} one or more spatial directions.}

\item{B}{\code{numeric} strength of magnetic field in Tesla.}

\item{particles}{\code{list} with mass, charge, and momentum distribution}
}
\value{
\code{data.frame}
}
\description{
Wrapper function for j0j0 to calculate a spectrum of j0j0 values
  for multiple frequencies and directions.
}
