% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j0j0_spectrum_no_future.R
\name{j0j0_spectrum_no_future}
\alias{j0j0_spectrum_no_future}
\title{j0j0_spectrum_no_future}
\usage{
j0j0_spectrum_no_future(
  k,
  phi,
  frequencies,
  directions = c("x", "y", "z"),
  B,
  particle
)
}
\arguments{
\item{k}{\code{numeric} length of the fluctuation wavevector.}

\item{phi}{\code{numeric} angle (in degrees) between the magnetic field and
the fluctuation wavevector}

\item{frequencies}{\code{numeric} fluctuation frequencies in Hz.}

\item{directions}{\code{character} one or more spatial directions.}

\item{B}{\code{numeric} strength of magnetic field in Tesla.}

\item{particle}{\code{list} with mass, charge, and momentum distribution}
}
\value{
\code{data.frame}
}
\description{
Wrapper function for j0j0 to calculate a spectrum of j0j0 values
  for multiple frequencies and directions. No parallelization.
}
