% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j0j0_integrand.R
\name{j0j0_sum_terms}
\alias{j0j0_sum_terms}
\title{j0j0_sum_terms}
\usage{
j0j0_sum_terms(l_values, p_perp, directions, k_perp, k_par, omega, omega_c,
  m, distribution)
}
\arguments{
\item{l_values}{\code{numeric} vector of l-values at which to evaluate terms.}

\item{p_perp}{\code{numeric} component of momentum vector perpendicular to
the magnetic field.}

\item{directions}{\code{character} vector with two spatial directions:
acombination of x, y, and z. e.g  xx, xy or yz.}

\item{k_perp}{\code{numeric} length of component of wavevector perpendicular
to the magnetic field.}

\item{k_par}{\code{numeric} length of component of wavevector parallel to the
magnetic field.}

\item{omega}{\code{numeric} angular cyclotron frequency of current
fluctuations.}

\item{omega_c}{\code{numeric} particle angular cyclotron frequency (omega_c =
qB/m).}

\item{m}{\code{numeric} particle mass in kg.}

\item{distribution}{\code{list} with the velocity distribution.}
}
\value{
\code{complex}
}
\description{
Function to calculate the terms of the sum in the integrand of
  eq. 25 of Bindslev 1996, JATP 58, 983.
}
