% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j0j0_element.R
\name{matrix_integrand}
\alias{matrix_integrand}
\title{matrix_integrand}
\usage{
matrix_integrand(p_perp, directions, k_perp, k_par, omega, omega_c, mass,
  distribution)
}
\arguments{
\item{p_perp}{\code{numeric} component of momentum vector perpendicular to
the magnetic field divided by p_scale. Should be near 1.}

\item{directions}{\code{character} vector with two spatial directions:
acombination of x, y, and z. e.g  xx, xy or yz.}

\item{k_perp}{\code{numeric} length of component of wavevector perpendicular
to the magnetic field.}

\item{k_par}{\code{numeric} length of component of wavevector parallel to the
magnetic field.}

\item{omega}{\code{numeric} angular cyclotron frequency of current
fluctuations.}

\item{omega_c}{\code{numeric} particle angular cyclotron frequency (omega_c =
qB/m).}

\item{mass}{\code{numeric} particle mass in kg.}

\item{distribution}{\code{list} with the velocity distribution.}
}
\value{
\code{matrix}
}
\description{
Vectorized wrapper for j0j0_integrand. Returns a matrix for use
  with the curbature package.
}
