% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maxwellian_example}
\alias{maxwellian_example}
\title{Examples of results for j0j0 for a sweep of frequency, directions and two
values of phi for a maxwellian distribution.}
\format{A tibble with 12,642 rows and 10 variables:
\describe{
  \item{k}{length of wave vector in 1/m}
  \item{phi}{angle between wave vector and magnetic field in degrees}
  \item{frequency }{in Hz}
  \item{B}{Magnetic field strength in Tesla}
  \item{particle}{particle name}
  \item{integration_method}{name of integration method}
  \item{directions}{string with direction names}
  \item{A}{mass number}
  \item{Z}{charge number}
  \item{j0j0}{result for j0j0}
  }}
\source{
File /data/raw/j0j0_examples
}
\usage{
maxwellian_example
}
\description{
Examples of results for j0j0 for a sweep of frequency, directions and two
values of phi for a maxwellian distribution.
}
\keyword{datasets}
