% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_maxwellian.R
\docType{data}
\name{maxwellian_expr}
\alias{maxwellian_expr}
\title{maxwellian_expr}
\format{
An object of class \code{expression} of length 1.
}
\usage{
maxwellian_expr
}
\description{
expression for a simple maxwellian momentum distribution in
  cylindrical coordinates
}
\keyword{datasets}
