% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_maxwellian.R
\name{maxwellian_func}
\alias{maxwellian_func}
\title{maxwellian_func}
\usage{
maxwellian_func(p_perp, p_par, n, p_term)
}
\arguments{
\item{p_perp}{\code{numeric} value of perpendicular momentum component}

\item{p_par}{\code{numeric} value of parallel momentum component}

\item{n}{\code{numeric} particle density.}

\item{p_term}{\code{numeric} Particle themal momentum}
}
\value{
\code{numeric} value of momentum distribution at (p_perp, p_par)
}
\description{
function to evaluate a simple maxwellian momentum distribution
}
