% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_maxwellian.R
\name{maxwellian_grad}
\alias{maxwellian_grad}
\title{maxwellian_grad}
\usage{
maxwellian_grad(p_perp, p_par, n, p_term)
}
\arguments{
\item{p_perp}{\code{numeric} value of perpendicular momentum component}

\item{p_par}{\code{numeric} value of parallel momentum component}

\item{n}{\code{numeric} particle density.}

\item{p_term}{\code{numeric} Particle themal momentum}
}
\value{
\code{list}
}
\description{
function to calculate the gradient of a  simple maxwellian
  momentum distribution with respect to parallel and perpencicular momentum
}
