% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxwellian_ring.R
\docType{data}
\name{maxwellian_ring_expr}
\alias{maxwellian_ring_expr}
\title{maxwellian_ring_expr}
\format{An object of class \code{expression} of length 1.}
\usage{
maxwellian_ring_expr
}
\description{
expression for a maxwellian ring distribution in
  cylindrical coordinates
}
\keyword{datasets}
