% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxwellian_ring.R
\name{maxwellian_ring_func}
\alias{maxwellian_ring_func}
\title{maxwellian_ring_func}
\usage{
maxwellian_ring_func(p_perp, p_par, n, p_width, p_rad, K)
}
\arguments{
\item{p_perp}{\code{numeric} value of perpendicular momentum component}

\item{p_par}{\code{numeric} value of parallel momentum component}

\item{n}{\code{numeric} particle density.}

\item{p_width}{\code{numeric} Width of ring momentum distribution}

\item{p_rad}{\code{numeric} Radius of ring momentum distribution}

\item{K}{\code{numeric} Integration constant}
}
\value{
\code{numeric} value of momentum distribution at (p_perp, p_par)
}
\description{
function to evaluate a simple maxwellian momentum distribution
}
