% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxwellian_ring.R
\name{maxwellian_ring_setup}
\alias{maxwellian_ring_setup}
\title{maxwellian_ring_setup}
\usage{
maxwellian_ring_setup(n, v_width, v_rad, A, name)
}
\arguments{
\item{n}{\code{numeric} Particle density.}

\item{v_width}{\code{numeric} Width of ring in terms of velocity.}

\item{v_rad}{\code{numeric} Radius of ring in terms of velocity.}

\item{A}{\code{numeric} Particle mass number}

\item{name}{\code{character} Name of distribution/particle}

\item{Z}{\code{numeric} Particle charge number}
}
\value{
\code{list} with momentum distribution setup
}
\description{
Function to return a maxwellian ring momentum distribution
}
