% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j0j0_integrand.R
\name{new_l_values}
\alias{new_l_values}
\title{new_l_values}
\usage{
new_l_values(max_term, l_values_done, sum_terms, nl)
}
\arguments{
\item{max_term}{\code{numeric} maximum term encountered so far}

\item{l_values_done}{\code{numeric} vector of l-values already done.}

\item{sum_terms}{\code{complex} sum terms encountered so far.}

\item{nl}{\code{numeric} number of l-values to expand with.}
}
\value{
\code{numeric}
}
\description{
construct a new vextor of l-values by expanding up and/or dowm
}
