% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_tools.R
\name{optimize_distribution_perp}
\alias{optimize_distribution_perp}
\title{optimize_distribution_perp}
\usage{
optimize_distribution_perp(distribution, p_perp)
}
\arguments{
\item{distribution}{\code{list} with the momentum distributions.}

\item{p_perp}{\code{numeric} value of perpendicular momentum component.}
}
\value{
\code{numeric} values of p_par and the distribution at the optimum
}
\description{
function to find the optimum of a momentum distribution for
  constant perpendicular momentum.
}
