% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilkie_sd.R
\name{plasma_parameter}
\alias{plasma_parameter}
\title{plasma_parameter}
\usage{
plasma_parameter(n_e, T_i_eV, T_e_eV)
}
\arguments{
\item{n_e}{\code{numeric} electron density.}

\item{T_i_eV}{\code{numeric} Ion temperaure in eV.}

\item{T_e_eV}{\code{numeric} Electron temperaure in eV.}
}
\value{
\code{numeric} plasma parameter
}
\description{
Function to calculate the plasma parameter (eq. 1.8 of Swanson
  2008)
}
