% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_distribution_data_frame.R
\name{plot_dist}
\alias{plot_dist}
\title{plot_dist}
\usage{
plot_dist(dist_df)
}
\arguments{
\item{dist_df}{\code{data.frame} output of
calculate_distribution_data_frame() containing the evaluated distribution}
}
\value{
\code{ggplot}
}
\description{
helper function to plot momentum distributions
}
