% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution.R
\name{plot_distribution}
\alias{plot_distribution}
\title{plot_distribution}
\usage{
plot_distribution(particles, v_par, v_perp, logscale = TRUE)
}
\arguments{
\item{particles}{\code{list} list of particles with distribution setups.}

\item{v_par}{\code{numeric} value of parallel velocity}

\item{v_perp}{\code{numeric} value of perpendicular velocity}

\item{logscale}{\code{logical} if TRUE sets log scale on y-axis}
}
\value{
\code{ggplot} 1D plot of distributions
}
\description{
function to plot velocity distributions
}
