% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_j0j0.R
\name{plot_j0j0}
\alias{plot_j0j0}
\title{plot_j0j0}
\usage{
plot_j0j0(j0j0_df, wrap_by, color_by)
}
\arguments{
\item{j0j0_df}{\code{data.frame} output of j0j0r::j0j0() containing j0j0
elements}

\item{wrap_by}{\code{charater} the plot will facet wrap by this variable in
j0j0_df}

\item{color_by}{\code{charater} lines will be colored according to this
variable in j0j0_df}
}
\value{
\code{ggplot}
}
\description{
helper function to plot the output of j0j0r::j0j0()
}
