% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_tools.R
\name{root_distribution_perp}
\alias{root_distribution_perp}
\title{root_distribution_perp}
\usage{
root_distribution_perp(distribution, p_perp, tol = 1e-10)
}
\arguments{
\item{distribution}{\code{list} with the momentum distributions.}

\item{p_perp}{\code{numeric} value of perpendicular momentum component.}

\item{tol}{\code{numeric} tolerance.}
}
\value{
\code{numeric} values of p_par and the distribution at the root
}
\description{
function to find a root (to within tol) of a momentum
  distribution for constant perpendicular momentum.
}
