% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilkie_sd.R
\docType{data}
\name{slowdown_expr_p}
\alias{slowdown_expr_p}
\title{slowdown_expr_p}
\format{
An object of class \code{expression} of length 1.
}
\usage{
slowdown_expr_p
}
\description{
expression for a slowdown momentum distribution,
  https://arxiv.org/abs/1808.01934v2, eq. 2.9, using only length of p
}
\keyword{datasets}
