% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilkie_sd.R
\name{slowdown_grad}
\alias{slowdown_grad}
\title{slowdown_grad}
\usage{
slowdown_grad(p_perp, p_par, n, K, p_c, p_b, b)
}
\arguments{
\item{p_perp}{\code{numeric} value of perpendicular momentum component}

\item{p_par}{\code{numeric} value of parallel momentum component}

\item{n}{\code{numeric} particle density.}

\item{K}{\code{numeric} integrations constant}

\item{p_c}{\code{numeric} critical momentum}

\item{p_b}{\code{numeric} birth momentum}

\item{b}{\code{numeric} transport parameter}
}
\value{
\code{list}
}
\description{
function to calculate the gradient of a slowdown
  momentum distribution with respect to parallel and perpencicular momentum
}
