% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilkie_sd.R
\name{thermal_velocity}
\alias{thermal_velocity}
\title{thermal_velocity}
\usage{
thermal_velocity(T_eV, m)
}
\arguments{
\item{T_eV}{\code{numeric} Particle temperaure in eV.}

\item{m}{\code{numeric} particle mass in kg.}
}
\value{
\code{numeric} thermal velocity
}
\description{
Function to calculate a particles thermal velocity
}
