% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilkie_sd.R
\name{wilkie_setup}
\alias{wilkie_setup}
\title{wilkie_setup}
\usage{
wilkie_setup(b, n, A, Z, birth_energy, n_e, T_e_eV, ions)
}
\arguments{
\item{b}{\code{numeric} b-parameter in Wilkie 2018 eq. 2.9. Quantifies
importance of transport. Suggested in the range 0 (no transport) to 10
(significant effect).}

\item{n}{\code{numeric} Particle density.}

\item{A}{\code{numeric} Particle mass number}

\item{Z}{\code{numeric} Particle charge number}

\item{birth_energy}{\code{numeric} Particle birth energy in eV.}

\item{n_e}{\code{numeric} Electron density.}

\item{T_e_eV}{\code{numeric} Electron temperature in eV.}

\item{ions}{\code{data frame} with information on all ion species, containing
columns "n" (ion density), "A" (ion mass number), and "Z" (ion charge
number)}
}
\value{
\code{list} with momentum distribution setup
}
\description{
Function to return a setup for a wilkie slowdown momentum
  distribution, based on Wilkie 2018, https://arxiv.org/abs/1808.01934v2, eq.
  2.9.
}
