% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bimaxwellian.R
\name{bimaxwellian_setup}
\alias{bimaxwellian_setup}
\title{bimaxwellian_setup}
\usage{
bimaxwellian_setup(n, T_eV_perp, T_eV_par, v_drift, A)
}
\arguments{
\item{n}{\code{numeric} Particle density.}

\item{T_eV_perp}{\code{numeric} Perpendicular temperaure in eV.}

\item{T_eV_par}{\code{numeric} Parallel temperaure in eV.}

\item{v_drift}{\code{numeric} drift velocity parallel to magnetic field in
m/s.}

\item{A}{\code{numeric} Particle mass number}
}
\value{
\code{list} with momentum distribution setup
}
\description{
Function to return a simple maxwellian momentum distribution
}
