% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_maxwellian.R
\name{maxwellian_setup}
\alias{maxwellian_setup}
\title{maxwellian_setup}
\usage{
maxwellian_setup(n, T_eV, A)
}
\arguments{
\item{n}{\code{numeric} Particle density.}

\item{T_eV}{\code{numeric} Temperaure in eV.}

\item{A}{\code{numeric} Particle mass number}
}
\value{
\code{list} with momentum distribution setup
}
\description{
Function to return a simple maxwellian momentum distribution
}
